Adding Talk Titles - brings up link question
adding talk titles, but wondering about URLs to the talks themselves. If we link to shared folder, the links will need changing when we move the talks out of the google shared folder. I think we need a folder here in gitHub. @hollyel @ekrimmel @laurarussell any other thoughts on this issue? (For now, when adding talk titles, I've removed the "link" syntax).